refactor: move drag code into a own hook
